requirements.txt
Flask==3.1.1
flask-cors==6.0.1
flask-session==0.5.0
passlib==1.7.4
SQLAlchemy==2.0.41
pydantic==2.11.7
uwsgi==2.0.31
pydantic-settings==2.10.1
python-multipart==0.0.20
python-dotenv==1.1.1
psycopg2-binary==2.9.10
dataclass-factory==2.16
python-logstash==0.4.6
SQLAlchemy-Utils==0.41.1
pyyaml==6.0.1
gspread==6.2.1
google-auth==2.48.0
loguru==0.7.3
